Skip the maturin wheel build when its script is unreachable (#211)
`test_maturin_wheel_build_snapshot` fails every scheduled mutmut run:
`sys.executable -m maturin build` dies with "Unable to find `maturin`
script" before any mutants are generated, aborting the whole baseline.

The `maturin` PyPI package resolves its own compiled binary by walking
each `sysconfig` scheme's `scripts` directory keyed off the running
interpreter's `sys.prefix`, not `sys.path` or `PATH`. Under mutmut's
`uv run --with mutmut==3.6.0` overlay, `sys.prefix` points at a
temporary environment layered on top of the project's own virtualenv:
the `maturin` module imports fine (via `sys.path`, so `cargo`/`rustc`
and `toolchain_available()` all report success), but the overlay never
received maturin's script, so the lookup comes up empty.

Add `maturin_script_locatable()` to `tests/helpers/maturin.py`,
mirroring maturin's own lookup, and skip the wheel-build test with a
precise reason when it reports the script unreachable. In a normal
virtualenv (CI, `build-wheels.yml`, local `uv run pytest`) `sys.prefix`
matches the install location, the script is found, and the real build
still runs — only the layered mutmut overlay skips.

Add unit tests pinning the new detector's true/false behaviour by
faking the `sysconfig` scheme lookup.

Rebasing onto main brought in `MaturinBuildError` (#144), which — combined
with the new `maturin_script_locatable()` detector — pushed
`tests/helpers/maturin.py` past pylint's 400-line module limit. Extract the
wheel-artifact snapshot parsers (`wheel_build_snapshot` and its private
helpers) into a new sibling module `tests/helpers/maturin_wheel.py`, and
re-export `wheel_build_snapshot` from `tests/helpers/maturin.py` so existing
import sites are unchanged.
